requirements.txt
streamlit==1.35.0
supabase==2.4.6
python-dotenv==1.0.1
nltk==3.8.1
pandas==2.2.2
